Release 2.5 on platformIO

+++ include/pickup.h
@@ -5,6 +5,6 @@
 #ifndef MEGAFM_CMAKE_PICKUP_H
 #define MEGAFM_CMAKE_PICKUP_H
 
-void doPickup(byte number,byte data);
+void doPickup(byte number, int data);
 
 #endif //MEGAFM_CMAKE_PICKUP_H
